Fixed the printer scaling - to a certain extent.
git-svn-id: http://trace2uml.tigris.org/svn/trace2uml/trunk@58 89f8ffb8-d1fc-0310-92ca-d9a8b2f27bcf

+++ Model/Package_LEOIFVGO/classes/class_ETHKHZIC/operations/operation_UZWYCSOE/code.cpp
@@ -4,7 +4,7 @@ if (!dc)
 
 wxCoord w, h;
 dc->GetSize(&w, &h);
-double scale = (double)(mySeqDataBase->GetGraphWidth() / w);
+double scale = (double)mySeqDataBase->GetGraphWidth() / (double)w;
 dc->SetUserScale(scale, scale);
 mySeqGraphTab->DrawOnDC(*dc);
